+++ src/core/abi/abi-type.ts
@@ -1,5 +1,5 @@
 import { ABIValue } from './abi-value'
-import { ValidationError } from './helpers'
+import { ValidationError } from './errors'
 import {
   ABIAddressType,
   ABIBoolType,
@@ -8,7 +8,7 @@ import {
   ABIStaticArrayType,
   ABIStringType,
   ABITupleType,
-  ABIUFixedType,
+  ABIUfixedType,
   ABIUintType,
   decodeAddress,
   decodeBool,
@@ -17,7 +17,7 @@ import {
   decodeStaticArray,
   decodeString,
   decodeTuple,
-  decodeUFixed,
+  decodeUfixed,
   decodeUint,
   dynamicArrayToString,
   encodeAddress,
@@ -27,7 +27,7 @@ import {
   encodeStaticArray,
   encodeString,
   encodeTuple,
-  encodeUFixed,
+  encodeUfixed,
   encodeUint,
   staticArrayToString,
   tupleToString,
@@ -49,7 +49,7 @@ export enum ABITypeName {
 
 export type ABIType =
   | ABIUintType
-  | ABIUFixedType
+  | ABIUfixedType
   | ABIAddressType
   | ABIBoolType
   | ABIByteType
@@ -63,7 +63,7 @@ export function encode(abiType: ABIType, value: ABIValue): Uint8Array {
     case ABITypeName.Uint:
       return encodeUint(abiType, value)
     case ABITypeName.Ufixed:
-      return encodeUFixed(abiType, value)
+      return encodeUfixed(abiType, value)
     case ABITypeName.Address:
       return encodeAddress(value)
     case ABITypeName.Bool:
@@ -86,7 +86,7 @@ export function decode(abiType: ABIType, bytes: Uint8Array): ABIValue {
     case ABITypeName.Uint:
       return decodeUint(abiType, bytes)
     case ABITypeName.Ufixed:
-      return decodeUFixed(abiType, bytes)
+      return decodeUfixed(abiType, bytes)
     case ABITypeName.Address:
       return decodeAddress(bytes)
     case ABITypeName.Bool:
@@ -104,7 +104,7 @@ export function decode(abiType: ABIType, bytes: Uint8Array): ABIValue {
   }
 }
 
-export function toString(abiType: ABIType): string {
+export function aBITypeToString(abiType: ABIType): string {
   switch (abiType.name) {
     case ABITypeName.Uint:
       return uintToString(abiType)
@@ -131,10 +131,9 @@ const staticArrayRegexp = /^([a-z\d[\](),]+)\[(0|[1-9][\d]*)]$/
 const ufixedRegexp = /^ufixed([1-9][\d]*)x([1-9][\d]*)$/
 export const MAX_LEN = 2 ** 16 - 1
 
-// TODO: determine naming convention for the mappers
-export function asABIType(str: string): ABIType {
+export function stringToABIType(str: string): ABIType {
   if (str.endsWith('[]')) {
-    const childType = asABIType(str.slice(0, str.length - 2))
+    const childType = stringToABIType(str.slice(0, str.length - 2))
     return {
       name: ABITypeName.DynamicArray,
       childType: childType,
@@ -153,7 +152,7 @@ export function asABIType(str: string): ABIType {
       throw new ValidationError(`array length exceeds limit ${MAX_LEN}`)
     }
     // Parse the array element type
-    const childType = asABIType(stringMatches[1])
+    const childType = stringToABIType(stringMatches[1])
 
     return {
       name: ABITypeName.StaticArray,
@@ -202,7 +201,7 @@ export function asABIType(str: string): ABIType {
     const tupleContent = parseTupleContent(str.slice(1, str.length - 1))
     const childTypes: ABIType[] = []
     for (let i = 0; i < tupleContent.length; i++) {
-      const ti = asABIType(tupleContent[i])
+      const ti = stringToABIType(tupleContent[i])
       childTypes.push(ti)
     }
 
@@ -255,74 +254,4 @@ function parseTupleContent(content: string): string[] {
   }
 
   return tupleStrings
-}
-
-export function isDynamic(type: ABIType): boolean {
-  switch (type.name) {
-    case ABITypeName.StaticArray:
-      return isDynamic(type.childType)
-    case ABITypeName.Tuple:
-      return type.childTypes.some((c) => isDynamic(c))
-    case ABITypeName.DynamicArray:
-    case ABITypeName.String:
-      return true
-    default:
-      return false
-  }
-}
-
-export function findBoolSequenceEnd(abiTypes: ABIType[], currentIndex: number): number {
-  let cursor = currentIndex
-  while (cursor < abiTypes.length) {
-    if (abiTypes[cursor].name === ABITypeName.Bool) {
-      if (cursor - currentIndex + 1 === 8 || cursor === abiTypes.length - 1) {
-        return cursor
-      }
-      cursor++
-    } else {
-      return cursor - 1
-    }
-  }
-  return cursor - 1
-}
-
-export function getSize(abiType: ABIType): number {
-  switch (abiType.name) {
-    case ABITypeName.Uint:
-      return Math.floor(abiType.bitSize / 8)
-    case ABITypeName.Ufixed:
-      return Math.floor(abiType.bitSize / 8)
-    case ABITypeName.Address:
-      return 32
-    case ABITypeName.Bool:
-      return 1
-    case ABITypeName.Byte:
-      return 1
-    case ABITypeName.StaticArray:
-      if (abiType.childType.name === ABITypeName.Bool) {
-        return Math.ceil(abiType.length / 8)
-      }
-      return getSize(abiType.childType) * abiType.length
-    case ABITypeName.Tuple: {
-      let size = 0
-      let i = 0
-      while (i < abiType.childTypes.length) {
-        const childType = abiType.childTypes[i]
-        if (childType.name === ABITypeName.Bool) {
-          const sequenceEndIndex = findBoolSequenceEnd(abiType.childTypes, i)
-          const boolCount = sequenceEndIndex - i + 1
-          size += Math.ceil(boolCount / 8)
-          i = sequenceEndIndex + 1
-        } else {
-          size += getSize(childType)
-          i++
-        }
-      }
-      return size
-    }
-    case ABITypeName.String:
-      throw new ValidationError(`Failed to get size, string is a dynamic type`)
-    case ABITypeName.DynamicArray:
-      throw new ValidationError(`Failed to get size, dynamic array is a dynamic type`)
-  }
 }
