--- src/core/abi/constants.ts
@@ -1,5 +1,0 @@
-// TODO: check for duplicates
-const BITS_PER_BYTE = 8
-const MAX_BIT_SIZE = 512
-const MAX_PRECISION = 160
-const ALGORAND_PUBLIC_KEY_BYTE_LENGTH = 32
