+++ src/core/abi/types/primitives/address.ts
@@ -1,6 +1,7 @@
+import { addressToBytes, bytesToAddress } from '../../../utils'
 import { ABITypeName } from '../../abi-type'
 import { ABIValue } from '../../abi-value'
-import { convertAddressToBytes, convertBytesToAddress, EncodingError } from '../../helpers'
+import { EncodingError } from '../../errors'
 
 export type ABIAddressType = {
   name: ABITypeName.Address
@@ -8,11 +9,11 @@ export type ABIAddressType = {
 
 export function encodeAddress(value: ABIValue): Uint8Array {
   if (typeof value === 'string') {
-    return convertAddressToBytes(value)
+    return addressToBytes(value)
   }
   throw new EncodingError(`Cannot encode value as address: ${value}`)
 }
 
 export function decodeAddress(bytes: Uint8Array): ABIValue {
-  return convertBytesToAddress(bytes)
+  return bytesToAddress(bytes)
 }
