+++ src/core/abi/types/primitives/uint.ts
@@ -1,7 +1,7 @@
 import { bigIntToBytes, bytesToBigInt } from '../../../utils'
 import { ABITypeName } from '../../abi-type'
 import { ABIValue } from '../../abi-value'
-import { ValidationError } from '../../helpers'
+import { ValidationError } from '../../errors'
 
 export type ABIUintType = {
   name: ABITypeName.Uint
