+++ src/core/constants.ts
@@ -1,0 +1,12 @@
+// TODO: check for duplicates
+export const STATIC_ARRAY_REGEX = /^(.+)\[(\d+)\]$/
+export const UFIXED_REGEX = /^ufixed(\d+)x(\d+)$/
+
+export const MALFORMED_ADDRESS_ERROR_MSG = 'address seems to be malformed'
+export const CHECKSUM_ADDRESS_ERROR_MSG = 'wrong checksum for address'
+export const ALGORAND_ADDRESS_LENGTH = 58
+export const ALGORAND_ADDRESS_BYTE_LENGTH = 36
+export const ALGORAND_CHECKSUM_BYTE_LENGTH = 4
+export const LENGTH_ENCODE_BYTE_SIZE = 2
+export const BOOL_TRUE_BYTE = 0x80
+export const BOOL_FALSE_BYTE = 0x00
