+++ src/core/utils.ts
@@ -1,3 +1,13 @@
+import base32 from 'hi-base32'
+import {
+  ALGORAND_ADDRESS_BYTE_LENGTH,
+  ALGORAND_ADDRESS_LENGTH,
+  ALGORAND_CHECKSUM_BYTE_LENGTH,
+  CHECKSUM_ADDRESS_ERROR_MSG,
+  MALFORMED_ADDRESS_ERROR_MSG,
+} from './constants'
+import * as nacl from './nacl/naclWrappers'
+
 /**
  * Check whether the environment is ReactNative
  * @returns True if ReactNative, false otherwise
@@ -62,4 +72,39 @@ export function bytesToBigInt(bytes: Uint8Array) {
     res = BigInt(Number(buf.getUint8(i))) + res * BigInt(256)
   }
   return res
+}
+
+// TODO: this will be replaced by the real logic when this branch is merged
+export function addressToBytes(address: string): Uint8Array {
+  if (typeof address !== 'string') throw new Error(`${MALFORMED_ADDRESS_ERROR_MSG}: expected string, got ${typeof address}, ${address}`)
+  if (address.length !== ALGORAND_ADDRESS_LENGTH)
+    throw new Error(`${MALFORMED_ADDRESS_ERROR_MSG}: expected length ${ALGORAND_ADDRESS_LENGTH}, got ${address.length}: ${address}`)
+
+  // try to decode
+  const decoded = base32.decode.asBytes(address)
+  // Sanity check
+  if (decoded.length !== ALGORAND_ADDRESS_BYTE_LENGTH)
+    throw new Error(`${MALFORMED_ADDRESS_ERROR_MSG}: expected byte length ${ALGORAND_ADDRESS_BYTE_LENGTH}, got ${decoded.length}`)
+
+  // Find publickey and checksum
+  const publicKey = new Uint8Array(decoded.slice(0, ALGORAND_ADDRESS_BYTE_LENGTH - ALGORAND_CHECKSUM_BYTE_LENGTH))
+  const checksum = new Uint8Array(decoded.slice(nacl.PUBLIC_KEY_LENGTH, ALGORAND_ADDRESS_BYTE_LENGTH))
+  const expectedChecksum = checksumFromPublicKey(publicKey)
+  // Check if the checksum and the address are equal
+  if (!arrayEqual(expectedChecksum, checksum)) throw new Error(CHECKSUM_ADDRESS_ERROR_MSG)
+
+  return publicKey
+}
+
+function checksumFromPublicKey(pk: Uint8Array): Uint8Array {
+  return Uint8Array.from(nacl.genericHash(pk).slice(nacl.HASH_BYTES_LENGTH - ALGORAND_CHECKSUM_BYTE_LENGTH, nacl.HASH_BYTES_LENGTH))
+}
+
+export function bytesToAddress(bytes: Uint8Array): string {
+  if (bytes.byteLength !== 32) {
+    throw new Error(`bytes must be 32 bytes long for an address`)
+  }
+  const checksumBytes = checksumFromPublicKey(bytes)
+  const addr = base32.encode(concatArrays(bytes, checksumBytes))
+  return addr.slice(0, ALGORAND_ADDRESS_LENGTH) // removing the extra '===='
 }
